# Samat_Check/FBA/portal/fba/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "FBA POS API",
    "version": "1.0"
  },
  "host": "FBAappuv01.fortebank.com:8082",
  "paths": {
...
    "/cities/count": {
      "get": {
        "description": "Получение количества городов",
        "produces": [
          "application/json"
        ],
        "tags": [
          "City"
        ],
        "operationId": "CityCount",
        "responses": {
          "200": {
            "description": "JSON с количеством городов",
            "schema": {
              "$ref": "#/definitions/Count"
            }
          },
          "500": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/city": {
      "get": {
        "description": "Получение списка городов",
        "produces": [
          "application/json"
        ],
        "tags": [
          "City"
        ],
        "operationId": "CityGET",
        "parameters": [
          {
            "type": "int",
            "description": "номер страницы пагинатора",
            "name": "page",
            "in": "query"
          },
          {
            "type": "int",
            "description": "количество записей на странице",
            "name": "size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "JSON массив городов",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/City"
              }
            }
          },
          "400": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "put": {
        "description": "Изменение города",
        "produces": [
          "application/json"
        ],
        "tags": [
          "City"
        ],
        "operationId": "CityPUT",
        "parameters": [
          {
            "description": "JSON для изменения модели города",
            "name": "bodyjson",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/City"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JSON ответ на изменение города",
            "schema": {
              "$ref": "#/definitions/City"
            }
          },
          "400": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "post": {
        "description": "Создание города",
        "produces": [
          "application/json"
        ],
        "tags": [
          "City"
        ],
        "operationId": "CityPOST",
        "parameters": [
          {
            "description": "JSON для создания города",
            "name": "bodyjson",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/City"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JSON созданная модель города",
            "schema": {
              "$ref": "#/definitions/City"
            }
          },
          "400": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/city/{id}": {
      "get": {
        "description": "Получение города по ID",
        "produces": [
          "application/json"
        ],
        "tags": [
          "City"
        ],
        "operationId": "CityGETByID",
        "parameters": [
          {
            "type": "string \u003cuuid4\u003e",
            "description": "идентификатор города",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "JSON модель города",
            "schema": {
              "$ref": "#/definitions/City"
            }
          },
          "400": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "delete": {
        "description": "Удаление города по ID",
        "produces": [
          "application/json"
        ],
        "tags": [
          "City"
        ],
        "operationId": "CityDELETE",
        "parameters": [
          {
            "type": "string \u003cuuid4\u003e",
            "description": "идентификатор города",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Успешное удаление",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
    "/worksession": {
      "get": {
        "description": "Получение списка рабочих смен",
        "produces": [
          "application/json"
        ],
        "tags": [
          "WorkSession"
        ],
        "operationId": "WorkSessionGET",
        "parameters": [
          {
            "type": "int",
            "description": "номер страницы пагинатора",
            "name": "page",
            "in": "query"
          },
          {
            "type": "int",
            "description": "количество записей на странице",
            "name": "size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "JSON массив рабочих смен",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/WorkSession"
              }
            }
          },
          "400": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "put": {
        "description": "Изменение рабочей смены",
        "produces": [
          "application/json"
        ],
        "tags": [
          "WorkSession"
        ],
        "operationId": "WorkSessionPUT",
        "parameters": [
          {
            "description": "JSON для изменения модели рабочей смены",
            "name": "bodyjson",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/WorkSession"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JSON ответ на изменение рабочей смены",
            "schema": {
              "$ref": "#/definitions/WorkSession"
            }
          },
          "400": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "post": {
        "description": "Создание рабочей смены",
        "produces": [
          "application/json"
        ],
        "tags": [
          "WorkSession"
        ],
        "operationId": "WorkSessionPOST",
        "parameters": [
          {
            "description": "JSON для создания рабочей смены",
            "name": "bodyjson",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/WorkSession"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JSON созданная модель рабочей смены",
            "schema": {
              "$ref": "#/definitions/WorkSession"
            }
          },
          "400": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/worksession/{id}": {
      "get": {
        "description": "Получение рабочей смены по ID",
        "produces": [
          "application/json"
        ],
        "tags": [
          "WorkSession"
        ],
        "operationId": "WorkSessionGETByID",
        "parameters": [
          {
            "type": "string \u003cuuid4\u003e",
            "description": "идентификатор рабочей смены",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "JSON модель рабочей смены",
            "schema": {
              "$ref": "#/definitions/WorkSession"
            }
          },
          "400": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "delete": {
        "description": "Удаление рабочей смены по ID",
        "produces": [
          "application/json"
        ],
        "tags": [
          "WorkSession"
        ],
        "operationId": "WorkSessionDELETE",
        "parameters": [
          {
            "type": "string \u003cuuid4\u003e",
            "description": "идентификатор рабочей смены",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Успешное удаление",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/worksessions/count": {
      "get": {
        "description": "Получение количества рабочих смен",
        "produces": [
          "application/json"
        ],
        "tags": [
          "WorkSession"
        ],
        "operationId": "WorkSessionCount",
        "responses": {
          "200": {
            "description": "JSON с количеством рабочих смен",
            "schema": {
              "$ref": "#/definitions/Count"
            }
          },
          "500": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "City": {
      "description": "City справочник городов",
      "type": "object",
      "properties": {
        "ID": {
          "type": "string",
          "format": "uuid4"
        },
        "Name": {
          "type": "string"
        }
      },
      "x-go-package": "fba/dbmodel"
    },
    "Company": {
      "description": "Company компании",
      "type": "object",
      "properties": {
        "Address": {
          "type": "string"
        },
        "Bin": {
          "type": "string"
        },
        "Contact": {
          "type": "string"
        },
        "ID": {
          "type": "string",
          "format": "uuid4"
        },
        "Name": {
          "type": "string"
        },
        "Staff": {
          "$ref": "#/definitions/Staff"
        },
        "StaffID": {
          "type": "string",
          "format": "uuid4"
        }
      },
      "x-go-package": "fba/dbmodel"
    },
    "Count": {
      "description": "modelRowsCount структура для количества записей под swagger",
      "type": "object",
      "properties": {
        "count": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Count"
        }
      },
      "x-go-name": "modelRowsCount",
      "x-go-package": "fba/model"
    },
...
    "Error": {
      "description": "Error Ошибка",
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "x-go-name": "Error"
        }
      },
      "x-go-package": "fba/model"
    },
...
    "Point": {
      "description": "Point точки продаж",
      "type": "object",
      "properties": {
        "Address": {
          "type": "string"
        },
        "City": {
          "$ref": "#/definitions/City"
        },
        "CityID": {
          "type": "string",
          "format": "uuid4"
        },
        "Company": {
          "$ref": "#/definitions/Company"
        },
        "CompanyID": {
          "type": "string",
          "format": "uuid4"
        },
        "ID": {
          "type": "string",
          "format": "uuid4"
        },
        "Name": {
          "type": "string"
        },
        "Staff": {
          "$ref": "#/definitions/Staff"
        },
        "StaffID": {
          "type": "string",
          "format": "uuid4"
        }
      },
      "x-go-package": "fba/dbmodel"
    },
...
    "Role": {
      "description": "Role права пользователей(staff)",
      "type": "object",
      "properties": {
        "ID": {
          "type": "string",
          "format": "uuid4"
        },
        "Name": {
          "type": "string"
        }
      },
      "x-go-package": "fba/dbmodel"
    },
    "Staff": {
      "description": "Staff пользователь",
      "type": "object",
      "properties": {
        "ID": {
          "type": "string",
          "format": "uuid4"
        },
        "Name": {
          "type": "string"
        },
        "ParentID": {
          "type": "string",
          "format": "uuid4"
        },
        "Role": {
          "$ref": "#/definitions/Role"
        },
        "RoleID": {
          "type": "string",
          "format": "uuid4"
        }
      },
      "x-go-package": "fba/dbmodel"
    },
...
    "WorkSession": {
      "description": "WorkSession смена",
      "type": "object",
      "properties": {
        "Endamount": {
          "type": "number",
          "format": "double"
        },
        "Enddate": {
          "type": "string",
          "format": "date-time"
        },
        "ID": {
          "type": "string",
          "format": "uuid4"
        },
        "Isactive": {
          "type": "boolean"
        },
        "Name": {
          "type": "string"
        },
        "Point": {
          "$ref": "#/definitions/Point"
        },
        "PointID": {
          "type": "string",
          "format": "uuid4"
        },
        "Staff": {
          "$ref": "#/definitions/Staff"
        },
        "StaffID": {
          "type": "string",
          "format": "uuid4"
        },
        "Startamount": {
          "type": "number",
          "format": "double"
        },
        "Startdate": {
          "type": "string",
          "format": "date-time"
        }
      },
      "x-go-package": "fba/dbmodel"
    }
  },
  "security": [
    {
      "api_key": []
    }
  ]
}